explicit type conversion to fix error on solaris
also vectorizing nth_prime func

+++ src/prime_count.cpp
@@ -1,12 +1,12 @@
 #include "primes.h"
 
-static const double prime_count_c = 30 * log(113) / 113;
+static const double prime_count_c = 30 * log((double) 113) / 113;
 
 //' @rdname prime_count
 //' @export
 // [[Rcpp::export]]
 int prime_count(int n, bool upper_bound) {
-  return (upper_bound ? prime_count_c : 1) * n / log(n);
+  return (upper_bound ? prime_count_c : 1) * n / log((double) n);
 }
 
 //' @rdname prime_count
@@ -14,5 +14,5 @@ int prime_count(int n, bool upper_bound) {
 // [[Rcpp::export]]
 int nth_prime_estimate(int n, bool upper_bound) {
   double c = upper_bound ? 0 : 1;
-  return n * (log(n * log(n)) - c);
+  return n * (log(n * log((double) n)) - c);
 }
